backend/requirements.txt
Django>=5.0
django-unfold
gunicorn
whitenoise
django-cors-headers
psycopg2-binary
python-dotenv
Pillow
